+++ NEWS.md
@@ -1,3 +1,5 @@
 # SpectraVis 0.2
 
 ## Changes in 0.2.1
+
+- Add `plotlySpectraMirror` function.